+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## X.X.X
+### PaymentSheet
+* [Fixed] Fixed a few design issues on visionOS.
+
 ## 23.20.0 2023-12-18
 ### PaymentSheet
 * [Added] Support for [card brand choice](https://stripe.com/docs/card-brand-choice). To set default preferred networks, use the new configuration option `PaymentSheet.Configuration.preferredNetworks`.